dalsza optymalizacja kodu
git-svn-id: svn://svn.nlp.ipipan.waw.pl/morfeusz/morfeusz@181 ff4e3ee1-f430-4e82-ade0-24591c43f1fd

+++ morfeusz/InflexionGraph.hpp
@@ -22,7 +22,7 @@ public:
 
     struct Edge {
         InterpretedChunk chunk;
-        unsigned int nextNode;
+        unsigned long nextNode;
     };
 
     void addPath(const std::vector<InterpretedChunk>& path, bool weak);
